+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/components/VpnSettingsSection.kt
@@ -5,6 +5,9 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.ArrowRight
 import androidx.compose.material.icons.outlined.Campaign
+import androidx.compose.material.icons.automirrored.outlined.CallSplit
+import androidx.compose.material.icons.outlined.AdminPanelSettings
+import androidx.compose.material.icons.outlined.Dns
 import androidx.compose.material.icons.outlined.Lan
 import androidx.compose.material.icons.outlined.Power
 import androidx.compose.material3.Icon
@@ -182,6 +185,35 @@ fun VpnSettingsSection(values: SettingsValues, actions: SettingsActions) {
 			)
 
 			if (values.showCensorshipSection) {
+			add(
+				SelectionItem(
+					leading = {
+						Icon(
+							Icons.AutoMirrored.Outlined.CallSplit,
+							stringResource(R.string.split_tunneling),
+							modifier = Modifier.size(iconSize.scaledWidth()),
+						)
+					},
+					trailing = {
+						Icon(
+							Icons.AutoMirrored.Outlined.ArrowRight,
+							stringResource(R.string.go),
+							modifier = Modifier.size(iconSize),
+						)
+					},
+					title = {
+						Text(
+							stringResource(R.string.split_tunneling),
+							style = MaterialTheme.typography.bodyLarge.copy(MaterialTheme.colorScheme.onSurface),
+						)
+					},
+					onClick = {
+						navController.navigate(Route.SplitTunneling)
+					},
+				),
+			)
+
+			if (showCensorshipSection) {
 				add(
 					SelectionItem(
 						leading = {